requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
oauthlib==3.0.2
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.2
